rename login -> auth
