+++ client/src/app/auth/main.vue
