+++ client/src/app/auth/routes.js
@@ -11,8 +11,8 @@ const Login = (resolve) => {
 
 export default [
   {
-    name: 'login.index',
-    path: '/login',
+    name: 'auth.index',
+    path: '/auth',
     component: Login,
     meta: { requiresAuth: false },
   },
